0-hash_table_create.c 1

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -10,7 +10,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *table;
-	size_t i;
+	unsigned long int i;
 
 	table = malloc(sizeof(hash_table_t));
 	if (!table)
